--- pkg/utils/policy_type.go
@@ -1,35 +1,0 @@
-/*
-    Copyright (C) 2020 Accurics, Inc.
-
-	Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-		http://www.apache.org/licenses/LICENSE-2.0
-
-	Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-*/
-
-package utils
-
-// CheckPolicyType checks if supplied policy type matches desired policy types
-func CheckPolicyType(rulePolicyType string, desiredPolicyTypes []string) bool {
-	normDesiredPolicyTypes := make(map[string]bool, len(desiredPolicyTypes))
-	normRulePolicyType := EnsureUpperCaseTrimmed(rulePolicyType)
-
-	for _, desiredPolicyType := range desiredPolicyTypes {
-		desiredPolicyType = EnsureUpperCaseTrimmed(desiredPolicyType)
-		normDesiredPolicyTypes[desiredPolicyType] = true
-	}
-
-	if _, ok := normDesiredPolicyTypes["ALL"]; ok {
-		return true
-	}
-
-	_, ok := normDesiredPolicyTypes[normRulePolicyType]
-	return ok
-}
